isl_multi_templ.c: extract out isl_multi_nan_templ.c
The isl_multi_*_involves_nan function is only useful
in case the base expressions can actually represent a NaN.
