+++ Makefile.am
@@ -446,6 +446,7 @@ EXTRA_DIST = \
 	isl_multi_gist.c \
 	isl_multi_hash.c \
 	isl_multi_intersect.c \
+	isl_multi_nan_templ.c \
 	isl_multi_param_templ.c \
 	print_templ.c \
 	print_templ_yaml.c \
