+++ include/isl/multi.h
@@ -38,8 +38,6 @@ __isl_null isl_multi_##BASE *isl_multi_##BASE##_free(			\
 isl_bool isl_multi_##BASE##_plain_is_equal(				\
 	__isl_keep isl_multi_##BASE *multi1,				\
 	__isl_keep isl_multi_##BASE *multi2);				\
-isl_bool isl_multi_##BASE##_involves_nan(				\
-	__isl_keep isl_multi_##BASE *multi);				\
 int isl_multi_##BASE##_find_dim_by_id(					\
 	__isl_keep isl_multi_##BASE *multi, enum isl_dim_type type,	\
 	__isl_keep isl_id *id);						\
@@ -160,6 +158,10 @@ __isl_give isl_multi_##BASE *isl_multi_##BASE##_sub(			\
 __isl_export								\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_neg(		 	\
 	__isl_take isl_multi_##BASE *multi);
+
+#define ISL_DECLARE_MULTI_NAN(BASE)					\
+isl_bool isl_multi_##BASE##_involves_nan(				\
+	__isl_keep isl_multi_##BASE *multi);
 
 #define ISL_DECLARE_MULTI_DIMS(BASE)					\
 isl_bool isl_multi_##BASE##_involves_dims(				\
