+++ include/isl/val.h
@@ -14,6 +14,7 @@ extern "C" {
 
 ISL_DECLARE_MULTI(val)
 ISL_DECLARE_MULTI_ARITH(val)
+ISL_DECLARE_MULTI_NAN(val)
 ISL_DECLARE_MULTI_DIMS(val)
 ISL_DECLARE_MULTI_WITH_DOMAIN(val)
