+++ isl_multi_templ.c
@@ -1341,26 +1341,6 @@ isl_bool FN(MULTI(BASE),plain_is_equal)(__isl_keep MULTI(BASE) *multi1,
 	return isl_bool_true;
 }
 
-/* Does "multi" involve any NaNs?
- */
-isl_bool FN(MULTI(BASE),involves_nan)(__isl_keep MULTI(BASE) *multi)
-{
-	int i;
-
-	if (!multi)
-		return isl_bool_error;
-	if (multi->n == 0)
-		return isl_bool_false;
-
-	for (i = 0; i < multi->n; ++i) {
-		isl_bool has_nan = FN(EL,involves_nan)(multi->u.p[i]);
-		if (has_nan < 0 || has_nan)
-			return has_nan;
-	}
-
-	return isl_bool_false;
-}
-
 #ifndef NO_DOMAIN
 /* Return the shared domain of the elements of "multi".
  *
